requirements.txt
appdirs==1.4.3
bitmap==0.0.6
numpy==1.12.1
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
requests==2.13.0
scipy==0.19.0
six==1.10.0
